+++ pybamm/input/parameters/lithium-ion/cathodes/LiNiCoO2_Ecker2015/nco_electrolyte_exchange_current_density_Ecker2015.py
@@ -1,4 +1,4 @@
-from pybamm import exp, constants, standard_parameters_lithium_ion
+from pybamm import exp, constants, Parameter
 
 
 def nco_electrolyte_exchange_current_density_Ecker2015(c_e, c_s_surf, T):
@@ -41,7 +41,7 @@ def nco_electrolyte_exchange_current_density_Ecker2015(c_e, c_s_surf, T):
     E_r = 4.36e4
     arrhenius = exp(-E_r / (constants.R * T)) * exp(E_r / (constants.R * 296.15))
 
-    c_p_max = standard_parameters_lithium_ion.c_p_max
+    c_p_max = Parameter("Maximum concentration in positive electrode [mol.m-3]")
 
     return (
         m_ref * arrhenius * c_e ** 0.5 * c_s_surf ** 0.5 * (c_p_max - c_s_surf) ** 0.5