+++ pybamm/models/submodels/particle/base_particle.py
@@ -23,19 +23,21 @@ class BaseParticle(pybamm.BaseSubModel):
 
     def _get_standard_concentration_variables(self, c_s, c_s_xav):
 
+        param = self.param
         c_s_surf = pybamm.surf(c_s)
-
         c_s_surf_av = pybamm.x_average(c_s_surf)
-        geo_param = pybamm.geometric_parameters
 
         if self.domain == "Negative":
             c_scale = self.param.c_n_max
-            active_volume = geo_param.a_n_dim * geo_param.R_n / 3
+            active_volume = param.epsilon_s_n
         elif self.domain == "Positive":
             c_scale = self.param.c_p_max
-            active_volume = geo_param.a_p_dim * geo_param.R_p / 3
+            active_volume = param.epsilon_s_p
+
+        c_s_rav = pybamm.r_average(c_s)
         c_s_av = pybamm.r_average(c_s_xav)
         c_s_av_vol = active_volume * c_s_av
+
         variables = {
             self.domain + " particle concentration": c_s,
             self.domain + " particle concentration [mol.m-3]": c_s * c_scale,
@@ -43,6 +45,10 @@ class BaseParticle(pybamm.BaseSubModel):
             "X-averaged "
             + self.domain.lower()
             + " particle concentration [mol.m-3]": c_s_xav * c_scale,
+            "R-averaged " + self.domain.lower() + " particle concentration": c_s_rav,
+            "R-averaged "
+            + self.domain.lower()
+            + " particle concentration [mol.m-3]": c_s_rav * c_scale,
             self.domain + " particle surface concentration": c_s_surf,
             self.domain
             + " particle surface concentration [mol.m-3]": c_scale * c_s_surf,