+++ tests/unit/test_expression_tree/test_unary_operators.py
@@ -397,6 +397,10 @@ class TestUnaryOperators(unittest.TestCase):
             pybamm.x_average(symbol_on_edges)
 
         # Particle domains
+        geo = pybamm.GeometricParameters()
+        l_n = geo.l_n
+        l_p = geo.l_p
+
         a = pybamm.Symbol(
             "a",
             domain="negative particle",
@@ -406,7 +410,7 @@ class TestUnaryOperators(unittest.TestCase):
         self.assertEqual(a.domain, ["negative particle"])
         self.assertIsInstance(av_a, pybamm.Division)
         self.assertIsInstance(av_a.children[0], pybamm.Integral)
-        self.assertEqual(av_a.children[1].id, pybamm.geometric_parameters.l_n.id)
+        self.assertEqual(av_a.children[1].id, l_n.id)
 
         a = pybamm.Symbol(
             "a",
@@ -417,7 +421,7 @@ class TestUnaryOperators(unittest.TestCase):
         self.assertEqual(a.domain, ["positive particle"])
         self.assertIsInstance(av_a, pybamm.Division)
         self.assertIsInstance(av_a.children[0], pybamm.Integral)
-        self.assertEqual(av_a.children[1].id, pybamm.geometric_parameters.l_p.id)
+        self.assertEqual(av_a.children[1].id, l_p.id)
 
     def test_r_average(self):
         a = pybamm.Scalar(1)
@@ -438,6 +442,15 @@ class TestUnaryOperators(unittest.TestCase):
             self.assertEqual(av_a.children[0].integration_variable[0].domain, r.domain)
             # electrode domains go to current collector when averaged
             self.assertEqual(av_a.domain, [])
+
+        # r-average of a symbol that is broadcast to x
+        # takes the average of the child then broadcasts it
+        a = pybamm.Scalar(1, domain="positive particle")
+        broad_a = pybamm.SecondaryBroadcast(a, "positive electrode")
+        average_broad_a = pybamm.r_average(broad_a)
+        self.assertIsInstance(average_broad_a, pybamm.PrimaryBroadcast)
+        self.assertEqual(average_broad_a.domain, ["positive electrode"])
+        self.assertEqual(average_broad_a.children[0].id, pybamm.r_average(a).id)
 
         # r-average of symbol that evaluates on edges raises error
         symbol_on_edges = pybamm.PrimaryBroadcastToEdges(1, "domain")