+++ tests/unit/test_models/test_submodels/test_current_collector/test_homogeneous_current_collector.py
@@ -9,7 +9,7 @@ import unittest
 
 class TestUniformModel(unittest.TestCase):
     def test_public_functions(self):
-        param = pybamm.standard_parameters_lithium_ion
+        param = pybamm.LithiumIonParameters()
 
         submodel = pybamm.current_collector.Uniform(param)
         variables = {