+++ tests/unit/test_models/test_submodels/test_electrolyte_conductivity/test_surface_form/test_leading_surface_form_conductivity.py
@@ -9,7 +9,7 @@ import unittest
 
 class TestLeadingOrderModel(unittest.TestCase):
     def test_public_functions(self):
-        param = pybamm.standard_parameters_lithium_ion
+        param = pybamm.LithiumIonParameters()
         a = pybamm.Scalar(0)
         a_n = pybamm.FullBroadcast(
             pybamm.Scalar(0), ["negative electrode"], "current collector"